+++ src/librustc/back/arm.rs
@@ -9,7 +9,7 @@
 // except according to those terms.
 
 use back::target_strs;
-use driver::session::sess_os_to_meta_os;
+use driver::config::cfg_os_to_meta_os;
 use metadata::loader::meta_section_name;
 use syntax::abi;
 
@@ -22,7 +22,7 @@ pub fn get_target_strs(target_triple: ~str, target_os: abi::Os) -> target_strs::
     return target_strs::t {
         module_asm: "".to_owned(),
 
-        meta_sect_name: meta_section_name(sess_os_to_meta_os(target_os)).to_owned(),
+        meta_sect_name: meta_section_name(cfg_os_to_meta_os(target_os)).to_owned(),
 
         data_layout: match target_os {
           abi::OsMacos => {
