+++ src/librustc/middle/entry.rs
@@ -9,7 +9,7 @@
 // except according to those terms.
 
 
-use driver::session;
+use driver::config;
 use driver::session::Session;
 use syntax::ast::{Crate, Name, NodeId, Item, ItemFn};
 use syntax::ast_map;
@@ -49,7 +49,7 @@ impl<'a> Visitor<()> for EntryContext<'a> {
 
 pub fn find_entry_point(session: &Session, krate: &Crate, ast_map: &ast_map::Map) {
     let any_exe = session.crate_types.borrow().iter().any(|ty| {
-        *ty == session::CrateTypeExecutable
+        *ty == config::CrateTypeExecutable
     });
     if !any_exe {
         // No need to find a main function
@@ -58,7 +58,7 @@ pub fn find_entry_point(session: &Session, krate: &Crate, ast_map: &ast_map::Map
 
     // If the user wants no main function at all, then stop here.
     if attr::contains_name(krate.attrs.as_slice(), "no_main") {
-        session.entry_type.set(Some(session::EntryNone));
+        session.entry_type.set(Some(config::EntryNone));
         return
     }
 
@@ -127,13 +127,13 @@ fn find_item(item: &Item, ctxt: &mut EntryContext) {
 fn configure_main(this: &mut EntryContext) {
     if this.start_fn.is_some() {
         *this.session.entry_fn.borrow_mut() = this.start_fn;
-        this.session.entry_type.set(Some(session::EntryStart));
+        this.session.entry_type.set(Some(config::EntryStart));
     } else if this.attr_main_fn.is_some() {
         *this.session.entry_fn.borrow_mut() = this.attr_main_fn;
-        this.session.entry_type.set(Some(session::EntryMain));
+        this.session.entry_type.set(Some(config::EntryMain));
     } else if this.main_fn.is_some() {
         *this.session.entry_fn.borrow_mut() = this.main_fn;
-        this.session.entry_type.set(Some(session::EntryMain));
+        this.session.entry_type.set(Some(config::EntryMain));
     } else {
         // No main function
         this.session.err("main function not found");
