+++ src/librustc/middle/trans/closure.rs
@@ -11,7 +11,7 @@
 
 use back::abi;
 use back::link::mangle_internal_name_by_path_and_seq;
-use driver::session::FullDebugInfo;
+use driver::config::FullDebugInfo;
 use lib::llvm::ValueRef;
 use middle::freevars;
 use middle::lang_items::ClosureExchangeMallocFnLangItem;
