+++ src/librustc/middle/typeck/mod.rs
@@ -61,7 +61,7 @@ independently:
 
 #![allow(non_camel_case_types)]
 
-use driver::session;
+use driver::config;
 
 use middle::resolve;
 use middle::ty;
@@ -414,9 +414,9 @@ fn check_for_entry_fn(ccx: &CrateCtxt) {
     let tcx = ccx.tcx;
     match *tcx.sess.entry_fn.borrow() {
         Some((id, sp)) => match tcx.sess.entry_type.get() {
-            Some(session::EntryMain) => check_main_fn_ty(ccx, id, sp),
-            Some(session::EntryStart) => check_start_fn_ty(ccx, id, sp),
-            Some(session::EntryNone) => {}
+            Some(config::EntryMain) => check_main_fn_ty(ccx, id, sp),
+            Some(config::EntryStart) => check_start_fn_ty(ccx, id, sp),
+            Some(config::EntryNone) => {}
             None => tcx.sess.bug("entry function without a type")
         },
         None => {}
